refactor(vsock-guest): split lib into focused modules
