+++ crates/vsock-guest/src/log.rs
@@ -1,0 +1,4 @@
+/// Log a message to stderr
+pub fn log(level: &str, msg: &str) {
+    eprintln!("[vsock-guest] [{level}] {msg}");
+}
